src/app/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "anthropic>=0.97.0",
    "fastapi>=0.136.1",
    "passlib>=1.7.4",
    "psycopg2-binary>=2.9.12",
    "pydantic-settings>=2.14.0",
    "python-jose>=3.5.0",
    "sqlalchemy>=2.0.49",
    "uvicorn>=0.46.0",
]
